# src/main/resources/api.yaml
openapi: 3.0.0
info:
  title: Renovation Man API
  description: Renovation Man form API
  version: 0.0.0
tags:
  - name: Question
    description: Everything about questions
  - name: Questionnaire
    description: Everything about forms

paths:
  /question:
    get:
      tags:
        - Question
      operationId: getQuestions
      summary: Returns all the Questions
      responses:
        '200':
          description: Return JSON Array of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  $ref: "#/components/schemas/Question"
  #        links:
  #          GetResponseById:
  #            operationId: getResponse

  #  /reponse/{reponseId}:
  #    get:
  #      tags:
  #        - Reponse
  #      operationId: getReponse
  #      summary: Return Reponse
  #      responses:
  #        '200':
  #          description: Reponse required as JSON
  #          content:
  #            application/json:
  #              schema:
  #                type: object
  #                $ref: "#/components/schemas/Reponse"


  /questionnaire:
    get:
      tags:
        - Questionnaire
      operationId: getQuestionnaireByEmail
      summary: get form by email
      parameters:
        - in: query
          name: email
          schema:
            type: string
          description: email to search
      responses:
        '200':
          description: list of questionnaires matching query
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  $ref: "#/components/schemas/Questionnaire"
    post:
      tags:
        - Questionnaire
      operationId: postQuestionnaire
      summary: submit question form
      requestBody:
        description: form to register
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Questionnaire"
      responses:
        '200':
          description: Return JSON submitted form
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/Questionnaire"
components:
  schemas:
    Question:
      type: object
      properties:
        id:
          type: integer
          format: int64
        intitule:
          type: string
        reponses:
          type: array
          items:
            $ref: "#/components/schemas/Reponse"
    Reponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        intitule:
          type: string
    Questionnaire:
      type: object
      properties:
        email:
          type: string
        reponses:
          type: array
          items:
            $ref: "#/components/schemas/Reponse"
          minItems: 3
          maxItems: 3